chore: rename tekton test helpers package and hopefully hound is ok with it.
